Crate fast_chemail

Package fast_chemail validates the email as it is defined in the HTML specification.

Enums

ParseError

Functions

is_valid_email

is_valid_email checks wheter an email address is valid.

parse_email

parse_email scans an email address to check wheter it is correct.